pyproject.toml
[tool.poetry]
name = "tcc"
version = "0.1.0"
description = ""
authors = ["Otavio <otavio.dantas@syngenta.com>"]

[tool.poetry.dependencies]
python = ">=3.10, <3.12"
pyspark = "^3.3.0"
numpy = "^1.23.3"
Pillow = "^9.2.0"
scipy = "^1.9.1"
matplotlib = "^3.6.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
